+++ calculate/bad_boy_stats.py
@@ -13,21 +13,21 @@ class BadBoyStats(object):
 
         self.rankings = {}
         # Load the scoring based on crime categories
-        with open("resources/crimecatsscoring.csv", mode="r", encoding="utf-8-sig") as infile:
+        with open("resources/crime_category_scoring.csv", mode="r", encoding="utf-8-sig") as infile:
             reader = csv.reader(infile)
             for rows in reader:
-                cat = rows[0].upper().strip()
-                if cat.startswith('"') and cat.endswith('"'):
-                    cat = cat[1:-1]
+                crime_category = rows[0].upper().strip()
+                if crime_category.startswith('"') and crime_category.endswith('"'):
+                    crime_category = crime_category[1:-1]
                 rank = int(rows[1])
-                self.rankings[cat] = rank
+                self.rankings[crime_category] = rank
 
         if not dev_bool:
             url = "https://www.usatoday.com/sports/nfl/arrests/"
             r = requests.get(url)
             data = r.text
             soup = BeautifulSoup(data, "html.parser")
-            self.badboydata = {}
+            self.bad_boy_data = {}
 
             # Scrape USA Today NFL crime database site
             for row in soup.findAll("tr"):
@@ -38,54 +38,58 @@ class BadBoyStats(object):
                     date = cells[0].text
                     pos = cells[3].text
                     case = cells[4].text.upper()
-                    tempcat = cells[5].text.upper()
+                    temp_category = cells[5].text.upper()
 
-                    for item in tempcat.strip().split(","):
-                        cat = item.strip()
-                        if cat in self.rankings:
-                            score = self.rankings.get(cat)
+                    for item in temp_category.strip().split(","):
+                        category = item.strip()
+                        if category in self.rankings:
+                            score = self.rankings.get(category)
                         else:
                             score = 0
-                            print("Crime ranking not found: %s\nAssigning score of 0." % cat)
+                            print("Crime ranking not found: %s\nAssigning score of 0." % category)
 
-                        if name not in self.badboydata:
-                            self.badboydata[name] = {
+                        if name not in self.bad_boy_data:
+                            self.bad_boy_data[name] = {
                                 "team": team,
                                 "date": date,
                                 "pos": pos,
                                 "case": case,
-                                "cat": cat,
+                                "category": category,
                                 "points": score
                             }
                         else:
-                            points = self.badboydata[name].get("points") + score
-                            self.badboydata[name]["points"] = points
+                            points = self.bad_boy_data[name].get("points") + score
+                            self.bad_boy_data[name]["points"] = points
 
                         if pos in ["CB", "LB", "DE", "DT", "S"]:
-                            if team not in self.badboydata:
-                                self.badboydata[team] = {
+                            if team not in self.bad_boy_data:
+                                self.bad_boy_data[team] = {
                                     "team": team,
                                     "date": date,
                                     "pos": pos,
                                     "case": case,
-                                    "cat": cat,
+                                    "category": category,
                                     "points": score
                                 }
                             else:
-                                points = self.badboydata[name].get("points") + score
-                                self.badboydata[name]["points"] = points
+                                points = self.bad_boy_data[name].get("points") + score
+                                self.bad_boy_data[name]["points"] = points
             if save_bool:
                 with open(league_test_dir +
                           "/" +
                           "bad_boy_data.pkl", "wb") as bb_out:
-                    pickle.dump(self.badboydata, bb_out, pickle.HIGHEST_PROTOCOL)
+                    pickle.dump(self.bad_boy_data, bb_out, pickle.HIGHEST_PROTOCOL)
         else:
             with open(league_test_dir +
                       "/" +
                       "bad_boy_data.pkl", "rb") as bb_in:
-                self.badboydata = pickle.load(bb_in)
+                self.bad_boy_data = pickle.load(bb_in)
 
-        print("{} bad boy records loaded".format(len(self.badboydata)))
+        if len(self.bad_boy_data) == 0:
+            print("NO bad boy records were loaded, please check your internet connection or the availability of "
+                  "'https://www.usatoday.com/sports/nfl/arrests/' and try generating a new report.")
+        else:
+            print("{} bad boy records loaded".format(len(self.bad_boy_data)))
 
     def check_bad_boy_status(self, name, team, pos):
         """ Looks up given player and returns number of 'bad boy' points based on scoring.
@@ -97,11 +101,11 @@ class BadBoyStats(object):
         :return: Integer number of bad boy points, crime recorded
         """
         total = 0
-        cat = ""
+        category = ""
         if pos == "DEF":
             name = team
-        if name in self.badboydata:
-            crime = self.badboydata.get(name)
+        if name in self.bad_boy_data:
+            crime = self.bad_boy_data.get(name)
             total = crime.get("points")
-            cat = crime.get("cat")
-        return total, cat
+            category = crime.get("category")
+        return total, category
