+++ config.ini
@@ -1,10 +1,10 @@
 [Fantasy_Football_Report_Settings]
-chosen_league_id = 169896
-league_of_emperors_id = 169896
-humangeo_id = 363515
+league_id = 169896
 report_directory_base_path = ./reports
-;accepts "default", or "#", where # is an integer between 1 and 17 defining the chosen week
+
+;chosen_week accepts "default", or "#", where # is an integer between 1 and 17 defining the chosen week
 chosen_week = default
+
 ;Multiple teams can be manually disquailfied from coaching efficiency eligibility (use comma separated list w/o spaces).
 ;Example: coaching_efficiency_disqualified_teams = Team One,Team Two
 coaching_efficiency_disqualified_teams =
@@ -15,6 +15,7 @@ root_folder_name = Fantasy_Football
 
 [Slack_Settings]
 post_to_slack = False
+slack_channel = fantasyfootball
 
 [OAuth_Settings]
 yql_cache_dir = ./authentication/oauth_token
