+++ resources/crime_category_scoring.csv
