+++ utils/slack_messenger.py
@@ -5,7 +5,8 @@ from slackclient import SlackClient
 
 
 class SlackMessenger(object):
-    def __init__(self):
+    def __init__(self, config):
+        self.config = config
         with open("./authentication/private.txt", "r") as auth_file:
             auth_data = auth_file.read().split("\n")
 
@@ -19,14 +20,14 @@ class SlackMessenger(object):
     def list_channels(self):
         return self.sc.api_call("channels.list")
 
-    def test_post_on_hg_slack(self, message):
+    def test_post_to_slack(self, message):
         print(self.sc.api_call("channels.info", channel="C0A56L9A4"))
         return self.sc.api_call(
             "chat.postMessage", channel="#apitests", text="<!here|here>:\n" + message,
             username="fantasy_football_report_bot", icon_emoji=":football:"
         )
 
-    def test_file_upload_on_hg_slack(self, upload_file):
+    def test_file_upload_to_slack(self, upload_file):
         print(self.sc.api_call("channels.info", channel="C0A56L9A4"))
         with open(upload_file, "r") as uf:
             file_to_upload = uf.read()
@@ -44,13 +45,13 @@ class SlackMessenger(object):
             print("fileUpload failed %s", response["error"])
         return response
 
-    def post_to_hg_fantasy_football_channel(self, message):
+    def post_to_selected_slack_channel(self, message):
         return self.sc.api_call(
-            "chat.postMessage", channel="#fantasyfootball", text="<!here|here>\n" + message,
-            username="fantasy_football_report_bot", icon_emoji=":football:"
+            "chat.postMessage", channel="#" + self.config.get("Slack_Settings", "slack_channel"),
+            text="<!here|here>\n" + message, username="fantasy_football_report_bot", icon_emoji=":football:"
         )
 
-    def upload_file_to_hg_fantasy_football_channel(self, upload_file, slack_channel_name):
+    def upload_file_to_selected_slack_channel(self, upload_file):
         file_name = upload_file.split("/")[-1]
         file_type = file_name.split(".")[-1]
         league_name = file_name.split(".")[-2].split("_")[0]
@@ -61,7 +62,7 @@ class SlackMessenger(object):
             file_to_upload = uf.read()
             response = self.sc.api_call(
                 "files.upload",
-                channels="#" + slack_channel_name,
+                channels="#" + self.config.get("Slack_Settings", "slack_channel"),
                 filename=file_name,
                 filetype=file_type,
                 file=file_to_upload,
